Moved scm extension to main direction and made package-private, test fixes
